# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "description": "\n *     ## Инструкция по использованию API:\n *     \n *     1. **Регистрация/Вход**: Сначала зарегистрируйтесь через `/register` или войдите через `/login`\n *     2. **Получите токен**: После успешной аутентификации вы получите токен в ответе\n *     3. **Авторизация**: Нажмите кнопку 'Authorize' и введите: `Bearer ваш_токен`\n *     4. **Тестирование**: Теперь можете тестировать защищенные endpoints\n *     \n *     ## Публичные endpoints (не требуют авторизации):\n *     - GET /tasks-all - Все задачи\n *     - GET /tasks/{id} - Конкретная задача\n *     - GET /tasks-user/{id} - Задачи пользователя\n *     \n *     ## Защищенные endpoints (требуют авторизации):\n *     - POST /tasks - Создать задачу\n *     - PATCH /tasks/{id} - Обновить задачу\n *     - DELETE /tasks/{id} - Удалить задачу\n *     - POST /logout - Выход\n *     - GET /user - Информация о пользователе",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу по ID (публичный)",
                "operationId": "73d05aa3a120408c3826224752091459",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу (требует аутентификации)",
                "operationId": "1c17cab818454ae38f23569d8ff0a610",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена или нет доступа"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Частичное обновление задачи (требует аутентификации)",
                "operationId": "6d59ece8989feb4cce53cf7c5d286c42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновленный заголовок"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Обновленное описание"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ожидает",
                                            "в работе",
                                            "на паузе",
                                            "выполнена"
                                        ],
                                        "example": "выполнена"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена или нет доступа"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks-user/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачи конкретного пользователя (публичный)",
                "operationId": "8cbed73aa7ca97f43bb5d5e32b1ae363",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу (требует аутентификации)",
                "operationId": "6c69ac25a6c9f4d455dd465a2dad05f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Новая задача"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание новой задачи"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ожидает",
                                            "в работе",
                                            "на паузе",
                                            "выполнена"
                                        ],
                                        "example": "ожидает"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Моя задача"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание задачи"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ожидает",
                            "в работе",
                            "на паузе",
                            "выполнена"
                        ],
                        "example": "ожидает"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Аутентификация через Sanctum Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}